+++ sdk/core/azure-core-http-vertx/src/main/java/com/azure/core/http/vertx/VertxHttpClient.java
@@ -46,18 +46,18 @@ import java.util.concurrent.TimeUnit;
 /**
  * {@link HttpClient} implementation for the Vert.x {@link io.vertx.core.http.HttpClient}.
  */
-class VertxAsyncHttpClient implements HttpClient {
-    private static final ClientLogger LOGGER = new ClientLogger(VertxAsyncHttpClient.class);
+class VertxHttpClient implements HttpClient {
+    private static final ClientLogger LOGGER = new ClientLogger(VertxHttpClient.class);
 
     final io.vertx.core.http.HttpClient client;
     private final Duration responseTimeout;
 
     /**
-     * Constructs a {@link VertxAsyncHttpClient}.
+     * Constructs a {@link VertxHttpClient}.
      *
      * @param client The Vert.x {@link io.vertx.core.http.HttpClient}
      */
-    VertxAsyncHttpClient(io.vertx.core.http.HttpClient client, Duration responseTimeout) {
+    VertxHttpClient(io.vertx.core.http.HttpClient client, Duration responseTimeout) {
         this.client = Objects.requireNonNull(client, "client cannot be null");
         this.responseTimeout = responseTimeout;
     }