+++ sdk/core/azure-core-http-vertx/src/test/java/com/azure/core/http/vertx/VertxHttpClientSingletonTests.java
@@ -17,22 +17,22 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 @Execution(ExecutionMode.SAME_THREAD) // because singleton http client, it can avoid race condition.
-public class VertxAsyncHttpClientSingletonTests {
+public class VertxHttpClientSingletonTests {
     private static final ConfigurationSource EMPTY_SOURCE = new TestConfigurationSource();
 
     @Test
     public void testSingletonClientInstanceCreation() {
         Configuration configuration = getConfiguration(true);
-        HttpClient client1 = new VertxAsyncHttpClientProvider(configuration).createInstance();
-        HttpClient client2 = new VertxAsyncHttpClientProvider(configuration).createInstance();
+        HttpClient client1 = new VertxHttpClientProvider(configuration).createInstance();
+        HttpClient client2 = new VertxHttpClientProvider(configuration).createInstance();
         assertEquals(client1, client2);
     }
 
     @Test
     public void testNonDefaultClientInstanceCreation() {
         Configuration configuration = getConfiguration(false);
-        HttpClient client1 = new VertxAsyncHttpClientProvider(configuration).createInstance();
-        HttpClient client2 = new VertxAsyncHttpClientProvider(configuration).createInstance();
+        HttpClient client1 = new VertxHttpClientProvider(configuration).createInstance();
+        HttpClient client2 = new VertxHttpClientProvider(configuration).createInstance();
         assertNotEquals(client1, client2);
     }
 
@@ -40,16 +40,16 @@ public class VertxAsyncHttpClientSingletonTests {
     public void testCustomizedClientInstanceCreationNotShared() {
         Configuration configuration = getConfiguration(false);
         HttpClientOptions clientOptions = new HttpClientOptions().setMaximumConnectionPoolSize(500);
-        HttpClient client1 = new VertxAsyncHttpClientProvider(configuration).createInstance(clientOptions);
-        HttpClient client2 = new VertxAsyncHttpClientProvider(configuration).createInstance(clientOptions);
+        HttpClient client1 = new VertxHttpClientProvider(configuration).createInstance(clientOptions);
+        HttpClient client2 = new VertxHttpClientProvider(configuration).createInstance(clientOptions);
         assertNotEquals(client1, client2);
     }
 
     @Test
     public void testNullHttpClientOptionsInstanceCreation() {
         Configuration configuration = getConfiguration(true);
-        HttpClient client1 = new VertxAsyncHttpClientProvider(configuration).createInstance(null);
-        HttpClient client2 = new VertxAsyncHttpClientProvider(configuration).createInstance(null);
+        HttpClient client1 = new VertxHttpClientProvider(configuration).createInstance(null);
+        HttpClient client2 = new VertxHttpClientProvider(configuration).createInstance(null);
         assertEquals(client1, client2);
     }